Refactor: separate innerDto, delete business logig in dto
